refac: moved SPL Transaction implementation func to segaggs.go
